requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
piper-tts
python-dotenv
